+++ src/cli/test.ts
@@ -1,7 +1,15 @@
 import { Runner } from './Runner';
 import { execSync } from 'child_process';
+import arg from 'arg';
+import { helpArgs, helpMessages } from './constants';
 
-export const test: Runner = async (args) => {
+export const test: Runner = async (args, ui) => {
+    const localArgs = arg(helpArgs);
+    if (localArgs['--help']) {
+        ui.write(helpMessages['test']);
+        return;
+    }
+  
     const testArgs = args._.slice(1); // first argument is `test`, need to get rid of it
     execSync(`npm test ${testArgs.join(' ')}`, { stdio: 'inherit' });
 };