+++ src/cli/run.ts
@@ -3,9 +3,14 @@ import { createNetworkProvider, argSpec } from '../network/createNetworkProvider
 import { findScripts, selectFile } from '../utils';
 import { UIProvider } from '../ui/UIProvider';
 import arg from 'arg';
+import { helpArgs, helpMessages } from './constants';
 
 export const run: Runner = async (args: Args, ui: UIProvider, context: RunnerContext) => {
-    const localArgs = arg(argSpec);
+    const localArgs = arg({ ...argSpec, ...helpArgs });
+    if (localArgs['--help']) {
+        ui.write(helpMessages['run']);
+        return;
+    }
 
     const { module: mod } = await selectFile(await findScripts(), {
         ui,