+++ src/types/file.ts
@@ -1,0 +1,4 @@
+export type File = {
+    name: string;
+    path: string;
+};