+++ src/utils/selection.utils.ts
@@ -3,21 +3,29 @@ import fs from 'fs/promises';
 import { UIProvider } from '../ui/UIProvider';
 import { SCRIPTS_DIR } from '../paths';
 import { COMPILE_END, getCompilablesDirectory } from '../compile/compile';
+import { File } from '../types/file';
 
-export const findCompiles = async (directory?: string) => {
+export const findCompiles = async (directory?: string): Promise<File[]> => {
     directory ??= await getCompilablesDirectory();
     const files = await fs.readdir(directory);
     const compilables = files.filter((file) => file.endsWith(COMPILE_END));
     return compilables.map((file) => ({
-        path: path.join(directory!, file),
+        path: path.join(directory, file),
         name: file.slice(0, file.length - COMPILE_END.length),
     }));
 };
 
-export const findScripts = async () =>
-    (await fs.readdir(SCRIPTS_DIR))
-        .filter((f) => f.endsWith('.ts'))
-        .map((f) => ({ path: path.join(SCRIPTS_DIR, f), name: path.parse(f).name }));
+export const findScripts = async (): Promise<File[]> => {
+    const dirents = await fs.readdir(SCRIPTS_DIR, { recursive: true, withFileTypes: true });
+    const scripts = dirents.filter((dirent) => dirent.isFile() && dirent.name.endsWith('.ts'));
+
+    return scripts
+        .map((script) => ({
+            name: path.join(script.path.slice(SCRIPTS_DIR.length), script.name),
+            path: path.join(SCRIPTS_DIR, script.path, script.name),
+        }))
+        .sort((a, b) => (a.name >= b.name ? 1 : -1));
+};
 
 export async function selectOption(
     options: { name: string; value: string }[],
@@ -39,14 +47,14 @@ export async function selectOption(
 }
 
 export async function selectFile(
-    files: { name: string; path: string }[],
+    files: File[],
     opts: {
         ui: UIProvider;
         hint?: string;
         import?: boolean;
     },
 ) {
-    let selected: { name: string; path: string };
+    let selected: File;
 
     if (opts.hint) {
         const found = files.find((f) => f.name.toLowerCase() === opts.hint?.toLowerCase());