+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for scripts in subdirectories, for example `scripts/counter/deploy.ts`
+- Added the ability to specify test files in `blueprint test` command, for example `blueprint test Counter` 
 
 ### Fixed